consistenthash/consistenthash_test.go
package consistenthash

import (
	"strconv"
	"testing"
)

func TestHashing(t *testing.T) {
	// 3个副本，自定义hash函数
	hash := New(3, func(key []byte) uint32 {
		i, _ := strconv.Atoi(string(key))
		return uint32(i)
	})

	// 创建的虚拟节点应为
	// 2, 4, 6, 12, 14, 16, 22, 24, 26
	hash.Add("2", "4", "6")

	// 测试用例
	testCases := map[string]string{
		"2":  "2",
		"11": "2",
		"23": "4",
		"27": "2",
	}

	for k, v := range testCases {
		if hash.Get(k) != v {
			t.Errorf("Asking for %s, should have yielded %s", k, v)
		}
	}

	// Adds 8, 18, 28
	hash.Add("8")

	// 27 should now map to 8.
	testCases["27"] = "8"

	for k, v := range testCases {
		if hash.Get(k) != v {
			t.Errorf("Asking for %s, should have yielded %s", k, v)
		}
	}

	// delete 8
	hash.Del("8")
	// 28 should now map to 2
	testCases["27"] = "2"

	for k, v := range testCases {
		if hash.Get(k) != v {
			t.Errorf("Asking for %s, should have yielded %s", k, v)
		}
	}
}
